--- a/makefile
+++ b/makefile
@@ -85,3 +85,17 @@ example/queens:
 	@echo ""
 	@./build/example/queens ${N} ${M}
 	@echo ""
+
+
+example/knights_tour: N := 5
+example/knights_tour:
+  # Build
+	@mkdir -p build/
+	@cd build/ && cmake -D CMAKE_BUILD_TYPE=Release ..
+
+	@cd build/ && make $(MAKE_FLAGS) knights_tour
+
+  # Run
+	@echo ""
+	@./build/example/knights_tour ${N} ${M}
+	@echo ""