--- a/src/adiar/adiar.cpp
+++ b/src/adiar/adiar.cpp
@@ -1,34 +1,34 @@
 #include "adiar.h"
 
-#include <tpie/tempname.h>
+#include <tpie/tpie.h>
+
+#include <adiar/internal/memory.h>
+
+#include <adiar/assert.h>
 
 namespace adiar
 {
+  const tpie::flags<tpie::subsystem> TPIE_SUBSYSTEMS =
+    tpie::MEMORY_MANAGER
+    | tpie::DEFAULT_LOGGING
+    | tpie::PROGRESS
+    // | tpie::JOB_MANAGER
+    // | tpie::CAPTURE_FRACTIONS
+    | tpie::STREAMS
+    | tpie::TEMPFILE
+    | tpie::FILE_MANAGER;
+
   void adiar_init(size_t memory_limit_bytes, std::string temp_dir)
   {
-    tpie::tpie_init();
+    tpie::tpie_init(TPIE_SUBSYSTEMS);
 
-    // Memory usage
-    set_limit(memory_limit_bytes);
-
-    // Temporary files
-    tpie::tempname::set_default_base_name("ADIAR");
-    tpie::tempname::set_default_extension("adiar"); // does this do anything?
-
-    if (temp_dir != "") {
-      tpie::tempname::set_default_path(temp_dir);
-    }
-  }
-
-  void set_limit(size_t memory_limit_bytes)
-  {
-    adiar_assert(memory_limit_bytes >= 128 * 1024 * 1024,
-                 "ADIAR requires at least 128 MiB of memory");
-    tpie::get_memory_manager().set_limit(memory_limit_bytes);
+    // Memory management
+    memory::init(temp_dir);
+    memory::set_limit(memory_limit_bytes);
   }
 
   void adiar_deinit()
   {
-    tpie::tpie_finish();
+    tpie::tpie_finish(TPIE_SUBSYSTEMS);
   }
 }