--- a/src/adiar/adiar.h
+++ b/src/adiar/adiar.h
@@ -34,6 +34,7 @@
 #include <adiar/zdd/build.h>
 #include <adiar/zdd/count.h>
 #include <adiar/zdd/pred.h>
+#include <adiar/zdd/project.h>
 #include <adiar/zdd/set_op.h>
 #include <adiar/zdd/zdd.h>