--- a/src/adiar/assert.h
+++ b/src/adiar/assert.h
@@ -54,6 +54,16 @@ namespace adiar
       }
   }
 
+#   define adiar_unreachable()                  \
+  __adiar_unreachable(__FILE__, __LINE__)
+
+  inline void __adiar_unreachable(const char* file, int line)
+  {
+    std::cerr << "Reached \"unreachable\" statement'\n"
+              << "Source:\t\t" << file << ", line " << line << "\n";
+    abort();
+  }
+
   // LCOV_EXCL_STOP
 }