--- a/src/adiar/bdd/apply.cpp
+++ b/src/adiar/bdd/apply.cpp
@@ -5,6 +5,7 @@
 #include <adiar/tuple.h>
 
 #include <adiar/internal/levelized_priority_queue.h>
+#include <adiar/internal/product_construction.h>
 
 #include <adiar/bdd/build.h>
 #include <adiar/bdd/negate.h>
@@ -14,61 +15,18 @@
 namespace adiar
 {
   //////////////////////////////////////////////////////////////////////////////
-  // Data structures
-  struct apply_tuple : tuple
+  // Apply policy for product construction.
+  class apply_prod_policy
   {
-    ptr_t source;
-  };
-
-  struct apply_tuple_data : tuple_data
-  {
-    ptr_t source;
-  };
-
-  typedef levelized_node_priority_queue<apply_tuple, tuple_label, tuple_fst_lt, std::less<>, 2> apply_priority_queue_1_t;
-  typedef tpie::priority_queue<apply_tuple_data, tuple_snd_lt> apply_priority_queue_2_t;
-
-  //////////////////////////////////////////////////////////////////////////////
-  // Helper functions
-  inline void apply_init_request(node_stream<> &in_nodes, node_t &v, label_t out_label,
-                                 ptr_t &low, ptr_t &high)
-  {
-    if (!is_sink(v) && label_of(v) == out_label) {
-      low = v.low;
-      high = v.high;
-
-      if (in_nodes.can_pull()) { v = in_nodes.pull(); }
-    } else {
-      low = high = v.uid;
+  public:
+    static bool is_same_file(const bdd &bdd_1, const bdd &bdd_2)
+    {
+      return bdd_1.file._file_ptr == bdd_2.file._file_ptr;
     }
-  }
 
-  inline void apply_resolve_request(apply_priority_queue_1_t &apply_pq_1,
-                                    arc_writer &aw,
-                                    const bool_op &op,
-                                    ptr_t source, ptr_t r1, ptr_t r2)
-  {
-    if (is_sink(r1) && is_sink(r2)) {
-      arc_t out_arc = { source, op(r1, r2) };
-      aw.unsafe_push_sink(out_arc);
-    } else if (is_sink(r1) && can_left_shortcut(op, r1)) {
-      arc_t out_arc = { source, op(r1, create_sink_ptr(true)) };
-      aw.unsafe_push_sink(out_arc);
-    } else if (is_sink(r2) && can_right_shortcut(op, r2)) {
-      arc_t out_arc = { source, op(create_sink_ptr(true), r2) };
-      aw.unsafe_push_sink(out_arc);
-    } else {
-      apply_pq_1.push({ r1, r2, source });
-    }
-  }
-
-  //////////////////////////////////////////////////////////////////////////////
-  __bdd bdd_apply(const bdd &bdd_1,
-                  const bdd &bdd_2,
-                  const bool_op &op)
-  {
-    // Resolve being given the same underlying input
-    if (bdd_1.file._file_ptr == bdd_2.file._file_ptr) {
+    static __bdd resolve_same_file(const bdd &bdd_1, const bdd &bdd_2,
+                                   const bool_op &op)
+    {
       ptr_t sink_1_F = create_sink_ptr(bdd_1.negate);
       ptr_t sink_2_F = create_sink_ptr(bdd_2.negate);
 
@@ -84,176 +42,58 @@ namespace adiar
       return op_F == sink_1_F ? bdd_1 : ~bdd_1;
     }
 
-    node_stream<> in_nodes_1(bdd_1);
-    node_stream<> in_nodes_2(bdd_2);
-
-    node_t v1 = in_nodes_1.pull();
-    node_t v2 = in_nodes_2.pull();
-
-    // Resolve sink shortcutting the result
-    if (is_sink(v1) && is_sink(v2)) {
-      ptr_t p = op(v1.uid, v2.uid);
-      return bdd_sink(value_of(p));
-    } else if (is_sink(v1)) {
-      if (can_left_shortcut(op, v1.uid)) {
-        ptr_t p =  op(v1.uid, create_sink_ptr(false));
+    static __bdd resolve_sink_root(const node_t &v1, const bdd& bdd_1,
+                                   const node_t &v2, const bdd& bdd_2,
+                                   const bool_op &op)
+    {
+      if (is_sink(v1) && is_sink(v2)) {
+        ptr_t p = op(v1.uid, v2.uid);
         return bdd_sink(value_of(p));
+      } else if (is_sink(v1)) {
+        if (can_left_shortcut(op, v1.uid)) {
+          ptr_t p =  op(v1.uid, create_sink_ptr(false));
+          return bdd_sink(value_of(p));
+        } else if (is_left_irrelevant(op, v1.uid)) {
+          return bdd_2;
+        } else { // if (is_left_negating(op, v1.uid)) {
+          return bdd_not(bdd_2);
+        }
+      } else { // if (is_sink(v2)) {
+        if (can_right_shortcut(op, v2.uid)) {
+          ptr_t p = op(create_sink_ptr(false), v2.uid);
+          return bdd_sink(value_of(p));
+        } else if (is_right_irrelevant(op, v2.uid)) {
+          return bdd_1;
+        } else { // if (is_right_negating(op, v2.uid)) {
+          return bdd_not(bdd_1);
+        }
       }
-      if (is_left_irrelevant(op, v1.uid)) {
-        return bdd_2;
-      }
-      if (is_left_negating(op, v1.uid)) {
-        return bdd_not(bdd_2);
-      }
-    } else if (is_sink(v2)) {
-      if (can_right_shortcut(op, v2.uid)) {
-        ptr_t p = op(create_sink_ptr(false), v2.uid);
-        return bdd_sink(value_of(p));
-      }
-      if (is_right_irrelevant(op, v2.uid)) {
-        return bdd_1;
-      }
-      if (is_right_negating(op, v2.uid)) {
-        return bdd_not(bdd_1);
-      }
+      adiar_unreachable();
     }
 
-    // Set-up for Apply Algorithm
-    arc_file out_arcs;
-    arc_writer aw(out_arcs);
-
-    tpie::memory_size_type available_memory = tpie::get_memory_manager().available();
-
-    // TODO: Do statistics on size of the 2 priority queues
-    apply_priority_queue_1_t apply_pq_1({bdd_1,bdd_2}, available_memory / 2);
-    apply_priority_queue_2_t apply_pq_2(calc_tpie_pq_factor(available_memory / 2));
-
-    // Process root and create initial recursion requests
-    label_t out_label = label_of(fst(v1.uid, v2.uid));
-    id_t out_id = 0;
-
-    ptr_t low1, low2, high1, high2;
-    apply_init_request(in_nodes_1, v1, out_label, low1, high1);
-    apply_init_request(in_nodes_2, v2, out_label, low2, high2);
-
-    // Shortcut the root
-    uid_t out_uid = create_node_uid(out_label, out_id++);
-    apply_resolve_request(apply_pq_1, aw, op, out_uid, low1, low2);
-    apply_resolve_request(apply_pq_1, aw, op, flag(out_uid), high1, high2);
-
-    // Process nodes in topological order of both BDDs
-    while (apply_pq_1.can_pull() || apply_pq_1.has_next_level() || !apply_pq_2.empty()) {
-      if (!apply_pq_1.can_pull() && apply_pq_2.empty()) {
-        aw.unsafe_push(create_meta(out_label, out_id));
-
-        apply_pq_1.setup_next_level();
-        out_label = apply_pq_1.current_level();
-        out_id = 0;
-      }
-
-      ptr_t source, t1, t2;
-      bool with_data = false;
-      ptr_t data_low = NIL, data_high = NIL;
-
-      // Merge requests from  apply_pq_1 or apply_pq_2
-      if (apply_pq_1.can_pull() && (apply_pq_2.empty() ||
-                                    fst(apply_pq_1.top()) < snd(apply_pq_2.top()))) {
-        source = apply_pq_1.top().source;
-        t1 = apply_pq_1.top().t1;
-        t2 = apply_pq_1.top().t2;
-
-        apply_pq_1.pop();
+    static void resolve_request(prod_priority_queue_1_t &prod_pq_1, arc_writer &aw,
+                                ptr_t /* t1 */, ptr_t /* t2 */, const bool_op &op,
+                                ptr_t source, ptr_t r1, ptr_t r2)
+    {
+      if (is_sink(r1) && is_sink(r2)) {
+        arc_t out_arc = { source, op(r1, r2) };
+        aw.unsafe_push_sink(out_arc);
+      } else if (is_sink(r1) && can_left_shortcut(op, r1)) {
+        arc_t out_arc = { source, op(r1, create_sink_ptr(true)) };
+        aw.unsafe_push_sink(out_arc);
+      } else if (is_sink(r2) && can_right_shortcut(op, r2)) {
+        arc_t out_arc = { source, op(create_sink_ptr(true), r2) };
+        aw.unsafe_push_sink(out_arc);
       } else {
-        source = apply_pq_2.top().source;
-        t1 = apply_pq_2.top().t1;
-        t2 = apply_pq_2.top().t2;
-
-        with_data = true;
-        data_low = apply_pq_2.top().data_low;
-        data_high = apply_pq_2.top().data_high;
-
-        apply_pq_2.pop();
-      }
-
-      adiar_invariant(is_sink(t1) || out_label <= label_of(t1),
-                      "Request should never level-wise be behind current position");
-      adiar_invariant(is_sink(t2) || out_label <= label_of(t2),
-                      "Request should never level-wise be behind current position");
-
-      // Seek request partially in stream
-      ptr_t t_seek = with_data ? snd(t1,t2) : fst(t1,t2);
-      while (v1.uid < t_seek && in_nodes_1.can_pull()) {
-        v1 = in_nodes_1.pull();
-      }
-      while (v2.uid < t_seek && in_nodes_2.can_pull()) {
-        v2 = in_nodes_2.pull();
-      }
-
-      // Forward information across the level
-
-      // TODO: We don't need to forward something, if there are no more nodes to
-      // pull from that stream.
-      if (is_node(t1) && is_node(t2) && label_of(t1) == label_of(t2)
-          && !with_data && (v1.uid != t1 || v2.uid != t2)) {
-        node_t v0 = v1.uid == t1 ? v1 : v2;
-
-        apply_pq_2.push({ t1, t2, v0.low, v0.high, source });
-
-        while (apply_pq_1.can_pull() && apply_pq_1.top().t1 == t1 && apply_pq_1.top().t2 == t2) {
-          source = apply_pq_1.pull().source;
-          apply_pq_2.push({ t1, t2, v0.low, v0.high, source });
-        }
-        continue;
-      }
-
-      // Resolve current node and recurse
-      // remember from above: ptr_t low1, low2, high1, high2;
-
-      if (is_sink(t1) || is_sink(t2) || label_of(t1) != label_of(t2)) {
-        if (t1 < t2) { // ==> label_of(t1) < label_of(t2) || is_sink(t2)
-          low1 = v1.low;
-          high1 = v1.high;
-          low2 = high2 = t2;
-        } else { // ==> label_of(t1) > label_of(t2) || is_sink(t1)
-          low1 = high1 = t1;
-          low2 = v2.low;
-          high2 = v2.high;
-        }
-      } else {
-        low1  = t1 < t_seek ? data_low  : v1.low;
-        high1 = t1 < t_seek ? data_high : v1.high;
-
-        low2  = t2 < t_seek ? data_low  : v2.low;
-        high2 = t2 < t_seek ? data_high : v2.high;
-      }
-
-      // Resolve request
-      adiar_debug(out_id < MAX_ID, "Has run out of ids");
-      out_uid = create_node_uid(out_label, out_id++);
-
-      apply_resolve_request(apply_pq_1, aw, op, out_uid, low1, low2);
-      apply_resolve_request(apply_pq_1, aw, op, flag(out_uid), high1, high2);
-
-      // Output ingoing arcs
-      while (true) {
-        arc_t out_arc = { source, out_uid };
-        aw.unsafe_push_node(out_arc);
-
-        if (apply_pq_1.can_pull() && apply_pq_1.top().t1 == t1 && apply_pq_1.top().t2 == t2) {
-          source = apply_pq_1.pull().source;
-        } else if (!apply_pq_2.empty() && apply_pq_2.top().t1 == t1 && apply_pq_2.top().t2 == t2) {
-          source = apply_pq_2.top().source;
-          apply_pq_2.pop();
-        } else {
-          break;
-        }
+        prod_pq_1.push({ r1, r2, source });
       }
     }
+  };
 
-    // Push the level of the very last iteration
-    aw.unsafe_push(create_meta(out_label, out_id));
-
-    return out_arcs;
+  //////////////////////////////////////////////////////////////////////////////
+  __bdd bdd_apply(const bdd &bdd_1, const bdd &bdd_2, const bool_op &op)
+  {
+    return product_construction<__bdd, bdd, apply_prod_policy>(bdd_1, bdd_2, op);
   }
 
   //////////////////////////////////////////////////////////////////////////////