--- a/src/adiar/bdd/apply.cpp
+++ b/src/adiar/bdd/apply.cpp
@@ -2,8 +2,9 @@
 
 #include <adiar/file_stream.h>
 #include <adiar/file_writer.h>
-#include <adiar/levelized_priority_queue.h>
 #include <adiar/tuple.h>
+
+#include <adiar/internal/levelized_priority_queue.h>
 
 #include <adiar/bdd/build.h>
 #include <adiar/bdd/negate.h>