… .github/workflows/examples.yml
… .github/workflows/static_analysis.yml
… README.md
… docs/bdd.md
… docs/core.md
… docs/core/data_types.md
… docs/examples.md
… docs/examples/queens.md
… docs/zdd.md
… example/CMakeLists.txt
… example/README.md
… example/examples_common.cpp
… example/knights_tour_bryant.cpp
… example/queens.cpp
… src/adiar/CMakeLists.txt
… src/adiar/bdd/build.cpp
… src/adiar/bdd/build.h
… src/adiar/bdd/count.cpp
… src/adiar/bdd/count.h
… src/adiar/bdd/evaluate.cpp
… src/adiar/bdd/evaluate.h
… src/adiar/bdd/if_then_else.cpp
… src/adiar/bdd/if_then_else.h
… src/adiar/bdd/quantify.cpp
… src/adiar/bdd/quantify.h
… src/adiar/bdd/restrict.cpp
… src/adiar/bdd/restrict.h
… src/adiar/data.cpp
… src/adiar/data.h
… src/adiar/file.cpp