@SSoelvsten 2021-09-20
Assuming I find a student programmer, then since the current version is in a not feature-complete but stable state, then I'll put this in _main_ as a **part 1**.